requirements.txt
streamlit>=1.20.0
pandas
joblib
scikit-learn
google-generativeai
